2025-09-26: Review add schema

+++ src/main/java/com/company/pgi/repository/profile/IProfileRepository.java
@@ -8,10 +8,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.company.pgi.model.Company;
 import com.company.pgi.model.Profile;
 
-public interface IProfileRepository extends JpaRepository<Profile, Integer> {
+public interface IProfileRepository extends JpaRepository<Profile, Long> {
 
     List<Profile> findProfileByCompany(Company company);
 
     Optional<Profile> findProfileById(Long id);
-
+    
 }
